more type refactor
